LULC_ Classification_Project/requirements.txt
# ============================================================
# TransUNet-RS — Python Dependencies
# ============================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
einops>=0.7.0

# Data & Image Processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
albumentations>=1.3.0
rasterio>=1.3.0
tifffile>=2023.7.0

# Training & Logging
tqdm>=4.65.0
tensorboard>=2.14.0
wandb>=0.15.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Evaluation & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API & Backend
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Utilities
click>=8.1.0
rich>=13.0.0
